changed the outputWrite() function to adapt PWM to the pump requirement
changed the outputWrite() function to adapt PWM to the pump requirement.
PWM values are intended to drive an OEG pump.
When R0 is forced, PWM is left unchanged so the test function can work.

added the testPWM menu that was used to determine min/max values needed for
OEG pump.
This menu may be removed as it is useless in normal operation.